pyproject.toml
[project]
name = "robot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "robotframework>=7.3.1",
    "robotframework-browser>=19.6.0",
    "robotframework-datadriver>=1.11.2",
]
